Updater and Cronjob integration

+++ ark_mod_manager.sh
@@ -24,7 +24,7 @@ ARK_MOD_ID=("525507438" "479295136" "632091170" "485964701" "558079412")
 ######## from here nothing change ########
 ##########################################
 
-CURRENT_VERSION="2.3"
+CURRENT_VERSION="2.4"
 ARK_APP_ID="346110"
 STEAM_MASTER_PATH="/home/$MASTERSERVER_USER/masterserver/steamCMD"
 STEAM_CMD_PATH="$STEAM_MASTER_PATH/steamcmd.sh"
@@ -96,13 +96,13 @@ MENU() {
 	echo
 	whiteMessage "1  -  Install a certain ModID"
 	whiteMessage "2  -  Install all ModIDs"
+	whiteMessage "3  -  Install Updater Script and Cronjob"
 	echo
-	whiteMessage "4  -  Update all installed ModIDs"
-#	whiteMessage "5  -  Install Updater Script + Cronjob"
-#	whiteMessage "6  -  Uninstall Updater Script"
+	whiteMessage "5  -  Update all installed ModIDs"
 	echo
-	whiteMessage "8  -  Uninstall a certain ModID"
-	whiteMessage "9  -  Uninstall all ModIDs"
+	whiteMessage "7  -  Uninstall a certain ModID"
+	whiteMessage "8  -  Uninstall all ModIDs"
+	whiteMessage "9  -  Uninstall Updater Script"
 	echo
 	whiteMessage "0  -  EXIT"
 	echo
@@ -116,23 +116,23 @@ MENU() {
 		2)
 			tput civis; MODE=INSTALLALL; INSTALL_ALL;;
 
-		4)
+		3)
+			tput civis; UPDATER_INSTALL;;
+
+		5)
 			tput civis; MODE=UPDATE; UPDATE;;
 
-#		5)
-#			tput civis; UPDATER_INSTALL;;
-#
-#		6)
-#			tput civis; UPDATER_UNINSTALL;;
-
-		8)
+		7)
 			tput civis; UNINSTALL;;
 
-		9)
+		8)
 			tput civis; UNINSTALL_ALL;;
 
+		9)
+			tput civis; UPDATER_UNINSTALL;;
+
 		0)
-			echo; clear; exit;;
+			tput cnorm; echo; clear; exit;;
 
 		*)
 			ERROR; MENU;;
@@ -229,31 +229,76 @@ UPDATE() {
 }
 
 UPDATER_INSTALL() {
-	if [ ! -f /etc/cron.d/ark_mod_updater ]; then
-		echo '30 1 * * * root /root/ark_mod_updater.sh >/dev/null 2>&1' > /etc/cron.d/ark_mod_updater
+	echo; echo
+	if [ -f "$MOD_LOG" ] || [ -f "$MOD_BACKUP_LOG" ]; then
+		yellowMessage "Check, is Cronjob already installed."
+		if [ ! -f /etc/cron.d/ark_mod_updater ]; then
+			echo '30 1 * * * root /root/ark_mod_updater.sh >/dev/null 2>&1' > /etc/cron.d/ark_mod_updater
 
-		if [ -f /etc/cron.d/ark_mod_updater ]; then
-			systemctl daemon-reload >/dev/null 2>&1 && service cron restart >/dev/null 2>&1
-			greenMessage "Updater Cron successfully installed."
+			if [ -f /etc/cron.d/ark_mod_updater ]; then
+				systemctl daemon-reload >/dev/null 2>&1 && service cron restart >/dev/null 2>&1
+				greenMessage "Updater Cron successfully installed."
+				sleep 3
+				echo
+			else
+				redMessage "Updater Cron installation failed!"
+				FINISHED
+			fi
 		else
-			redMessage "Updater Cron installation failed!"
+			greenMessage "Updater Cron already installed."
+			sleep 3
+			echo
+		fi
+
+		yellowMessage "Check, is a old Updater Script installed."
+		if [ -f /root/ark_mod_updater.sh ]; then
+			rm -rf /root/ark_mod_updater.sh
+			redMessage "old Updater Script found and removed."
+			sleep 3
+			echo
+		else
+			greenMessage "No old Script found."
+			sleep 3
+			echo
+		fi
+
+		if [ ! -f /root/ark_mod_updater.sh ] && [ -f /etc/cron.d/ark_mod_updater ]; then
+			yellowMessage "Downloading current Updater Script from Github"
+
+			wget --no-check-certificate https://raw.githubusercontent.com/Lacrimosa99/Easy-WI-ARK-Mod-Manager/master/ark_mod_updater.sh >/dev/null 2>&1
+			chmod 700 /root/ark_mod_updater.sh >/dev/null 2>&1
+
+			sed -i "s/unknown_user/$MASTERSERVER_USER/" /root/ark_mod_updater.sh
+
+			if [ ! "$STEAM_USERNAME" = "" ] && [ ! "$STEAM_PASSWD" = "" ]; then
+				sed -i "s/STEAM_USERNAME=/STEAM_USERNAME=\"$STEAM_USERNAME\"/" /root/ark_mod_updater.sh
+				sed -i "s/STEAM_PASSWD=/STEAM_PASSWD=\"$STEAM_PASSWD\"/" /root/ark_mod_updater.sh
+			fi
+
+			if [ ! "$EMAIL_TO" = "" ]; then
+				sed -i "s/EMAIL_TO=/EMAIL_TO=\"$EMAIL_TO\"/" /root/ark_mod_updater.sh
+			fi
+
+			greenMessage "Updater successfully installed."
+		else
+			redMessage "Updater installation failed!"
+			redMessage "Cron and Updater will be removed!"
+			sleep 3
+			UPDATER_UNINSTALL
 		fi
 	else
-		redMessage "Updater Cron already installed."
+		redMessage "Please install a Mod first, before you install the Updater!"
 	fi
-
-	if [ ! -f /root/ark_mod_updater.sh ] && [ -f /etc/cron.d/ark_mod_updater ]; then
-
-	else
-		redMessage "Updater installation failed!"
-	fi
+	FINISHED
 }
 
 UPDATER_UNINSTALL() {
+	echo; echo
 	if [ -f /etc/cron.d/ark_mod_updater ]; then
 		rm -rf /etc/cron.d/ark_mod_updater
 
 		if [ ! -f /etc/cron.d/ark_mod_updater ]; then
+			systemctl daemon-reload >/dev/null 2>&1 && service cron restart >/dev/null 2>&1
 			greenMessage "Updater Cron successfully uninstalled."
 		else
 			redMessage "Updater Cron uninstalling failed!"
